Rename channel-state folder to channel-store folder
