+++ packages/wallet/src/redux/channel-store/challenging/reducer.ts
