+++ packages/wallet/src/redux/channel-store/challenging/state.ts
