+++ packages/wallet/src/redux/channel-store/closing/__tests__/closing.test.ts
