+++ packages/wallet/src/redux/channel-store/closing/reducer.ts
