+++ packages/wallet/src/redux/channel-store/funding/__tests__/funding.test.ts
