+++ packages/wallet/src/redux/channel-store/funding/reducer.ts
