+++ packages/wallet/src/redux/channel-store/opening/reducer.ts
