+++ packages/wallet/src/redux/protocols/challenging/__tests__/scenarios.ts
@@ -2,7 +2,7 @@ import * as states from '../states';
 import * as actions from '../actions';
 import * as tsScenarios from '../../transaction-submission/__tests__';
 import { setChannel, EMPTY_SHARED_DATA } from '../../../state';
-import { ChannelStatus, waitForPreFundSetup, waitForUpdate } from '../../../channel-state/state';
+import { ChannelStatus, waitForPreFundSetup, waitForUpdate } from '../../../channel-store/state';
 import * as channelScenarios from '../../../__tests__/test-scenarios';
 
 type Reason = states.FailureReason;
