+++ packages/wallet/src/redux/protocols/concluding/__tests__/scenarios.ts
@@ -9,7 +9,7 @@ const { channelId, libraryAddress, channelNonce, participants } = channelScenari
 const channel = { channelId, libraryAddress, channelNonce, participants };
 const { asAddress: address, asPrivateKey: privateKey } = channelScenarios;
 const participant = { address, privateKey, ourIndex: 0 };
-import { ChannelStatus, waitForUpdate } from '../../../channel-state/state';
+import { ChannelStatus, waitForUpdate } from '../../../channel-store/state';
 const channelDefaults = { ...channel, ...participant };
 import { setChannel, EMPTY_SHARED_DATA } from '../../../state';
