+++ packages/wallet/src/redux/protocols/indirect-funding/__tests__/reducer.test.ts
@@ -7,7 +7,7 @@ import * as playerB from '../player-b/reducer';
 import * as scenarios from '../../../__tests__/test-scenarios';
 import { PlayerIndex } from '../../../types';
 import { emptyDisplayOutboxState } from '../../../outbox/state';
-import { emptyChannelState } from '../../../channel-state/state';
+import { emptyChannelState } from '../../../channel-store/state';
 const { channelId } = scenarios;
 const emptySharedData = {
   outboxState: emptyDisplayOutboxState(),
