+++ packages/wallet/src/redux/protocols/indirect-funding/player-a/__tests__/reducer.test.ts
@@ -2,7 +2,7 @@ import { MESSAGE_RELAY_REQUESTED } from 'magmo-wallet-client';
 import { addHex } from '../../../../../utils/hex-utils';
 import * as SigningUtil from '../../../../../utils/signing-utils';
 import * as actions from '../../../../actions';
-import * as channelStates from '../../../../channel-state/state';
+import * as channelStates from '../../../../channel-store/state';
 import { ProtocolStateWithSharedData } from '../../../../protocols';
 import { itSendsThisMessage, itTransitionsToChannelStateType } from '../../../../__tests__/helpers';
 import * as testScenarios from '../../../../__tests__/test-scenarios';
