+++ packages/wallet/src/redux/protocols/indirect-funding/player-a/__tests__/scenarios.ts
@@ -1,6 +1,6 @@
 import { PlayerIndex } from 'magmo-wallet-client/lib/wallet-instructions';
 import { emptyDisplayOutboxState } from '../../../../outbox/state';
-import * as channelStates from '../../../../channel-state/state';
+import * as channelStates from '../../../../channel-store/state';
 import { ProtocolStateWithSharedData } from '../../../../protocols';
 import * as testScenarios from '../../../../__tests__/test-scenarios';
 import {
