+++ packages/wallet/src/redux/protocols/indirect-funding/player-b/reducer.ts
@@ -23,7 +23,7 @@ import {
 } from '../reducer-helpers';
 import { ProtocolStateWithSharedData } from '../../';
 import { FundingAction, isDirectFundingAction } from '../../direct-funding/actions';
-import * as channelState from '../../../channel-state/state';
+import * as channelState from '../../../channel-store/state';
 import { Commitment } from 'fmg-core/lib/commitment';
 import { composePreFundCommitment } from '../../../../utils/commitment-utils';
 import { PlayerIndex } from '../../../types';
